growops: squash a -Wsign-conversion warning
By construction, `newopn` can only ever be non-negative here. Though in the long
term a more sweeping change should be done to count items in `size_t` variables
pervasively.

+++ lib/pathplan/route.c
@@ -418,7 +418,7 @@ static int growops(int newopn)
 {
     if (newopn <= opn)
 	return 0;
-    if (!(ops = realloc(ops, POINTSIZE * newopn))) {
+    if (!(ops = realloc(ops, POINTSIZE * (size_t)newopn))) {
 	return -1;
     }
     opn = newopn;
